+++ ElectronicObserver.KancolleApi.Types/Models/ApiHougeki1.cs
@@ -25,7 +25,7 @@ public class ApiHougeki1
 	[JsonPropertyName("api_damage")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
 	[Required]
-	public List<List<int>> ApiDamage { get; set; } = new();
+	public List<List<double>> ApiDamage { get; set; } = new();
 
 	[JsonPropertyName("api_df_list")]
 	[JsonIgnore(Condition = JsonIgnoreCondition.Never)]
